+++ modules/base/timer.c
@@ -117,7 +117,7 @@ static unsigned int clone_open(struct system_node *self)
     system_group_add(&snode[index].base, &snode[index].control.node);
     system_group_add(&snode[index].base, &snode[index].sleep.node);
 
-    return (unsigned int)&snode[index].control;
+    return (unsigned long)&snode[index].control;
 
 }
